recommendation-service/requirements.txt
fastapi
uvicorn
requests
numpy
scikit-learn
python-dotenv
sqlalchemy
psycopg2
python-multipart
python-jose[cryptography]